srcs/requirements/backend/requirements.txt
anyio==3.7.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==75.6.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
watchfiles==1.0.4
zope.interface==7.2
